Use dpcpp toolchain to build kernels (#983)
* Use dpcpp toolchain to build kernels

This patch introduces script to build our kernels
using tools from dpcpp toolchain.
It should be easier to maintain compared to the existing approach
with a dedicated opencl compiler:
- Reduce the number of dependencies
- Regular updates

Also the patch fixes build errors from the new tools:
usage of size_t as a kernel parameter(prohibited by OpenCL)
is replaced with ulong and some additional clean up to make
several kernels buildable.

All the kernels have been rebuild using the new tools.

Note: the script is only a temporary solution, eventually,
we'll build kernels through cmake.

+++ src/kernels/a2a_allreduce.cl
@@ -70,7 +70,7 @@
 
 __kernel void allreduce_execution_float32(int rank_id,
                                           int comm_size,
-                                          size_t elems_count,
+                                          ulong elems_count,
                                           const __global float* input_buffer,
                                           __global float* output_buffer,
                                           __global a2a_gpu_comm_data_float32* comm_matrix) {
